+++ CHANGELOG.md
@@ -30,6 +30,7 @@ We've also changed the following things in Server Manager itself:
 * Blue Flag detection! Live Timings now displays a blue flag indicator for cars that are being approached by drivers who have completed more laps in a race session.
 * DRS detection! Live Timings now displays an icon when a driver activates DRS.
 * Live Timings will now accurately show driver order even mid lap in Race Sessions. If one driver overtakes another, you should see their positions change in the Live Timings table!
+* Live Timings will now show the number of laps a driver has completed since their last pit stop.
 * Driver Swaps: you can now separate names by ';', in the same order as the GUIDs, so e.g. Driver Name 1;Driver Name 2;Driver Name 3 would match GUID1;GUID2;GUID3, and drivers names will then show correctly in Live Timings and results files.
 * Spectator Password - you can now connect to the server using a Spectator Password, and access commands to help spectating events!